requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
fastapi==0.112.2
fonttools==4.59.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.5
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psutil==7.0.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.38.6
threadpoolctl==3.6.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.30.6
watchfiles==1.1.0
websockets==15.0.1
openai>=1.40.0
jinja2>=3.1.6
pandas>=2.3.1
croniter>=1.3.0

# OpenTelemetry and observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-exporter-otlp-proto-grpc>=1.20.0

# Payment processing
redis==5.0.1
